src/requirements.txt
python-telegram-bot
requests
langdetect
google-cloud-texttospeech
google-cloud-translate
